+++ hw/mips/mips_ls2k.c
@@ -77,7 +77,7 @@ static qemu_irq *ls2k_irq,*ls2k_irq1;
 /* i8254 PIT is attached to the IRQ0 at PIC i8259 */
 
 static struct _loaderparams {
-    int ram_size;
+    uint64_t ram_size;
     const char *kernel_filename;
     const char *kernel_cmdline;
     const char *initrd_filename;
